+++ eslint/babel-eslint-tests/package.json
@@ -4,7 +4,8 @@
   "description": "Tests for babel/eslint-* packages",
   "license": "MIT",
   "private": true,
-  "devDependencies": {
+  "dependencies": {
+    "@babel/eslint-parser": "^7.11.0",
     "dedent": "^0.7.0",
     "eslint": "^7.5.0",
     "eslint-plugin-import": "^2.22.0"
